Move docs from knative/serving to knative/pkg (#102)
* Move docs from knative/serving to knative/pkg

While working on knative/build-pipeline#16
I started trying to build some integration tests, and I wanted to use
the code which had been moved to knative/pkg, but it turned out the docs
hadn't been ported with the code, so I am moving the docs relevant to
knative/pkg/test to that repo.

Removing docs from knative/serving in https://github.com/knative/serving/pull/2097.

* Fix knative/serving specific docs

Fixed copy-pasted references to `knative/serving/test` which should be
`knative/pkg/test`.

Fixed examples which were specific to `knative/serving`; in the case of
the command arg examples, removed the parts which are specific to those
e2e tests (though this would make them harder to use by copy and paste,
but since the libs are in a common repo now anyway that's probably
okay).

Tried to make the metric legend easier to read.