Make "Severity: Error" conditions less scary. (#255)
This restores the appearance of "terminal" conditions to what is was prior to the separation of "terminal" and "non-terminal" conditions.

Terminal conditions with an "Error" severity will simply omit the field.

This keeps coming up, and came to a head when both Evan and Ville bugged me about it in the same day.

Fixes: https://github.com/knative/serving/issues/2507