Adding default zipkin tracing support to SpoofClient
Existing support for Zipkin tracing pushes the
responsibility of logging zipkin traces to
individual tests. This change moves the responsibility
to SpoofClient, so that all tests get it by default.
Traces are logged only for requests that have
encountered Server errors i.e if the
http.Response.StatusCode between 500 to 599.

For SpoofingClient.Poll calls we only log trace of the
final SpoofingClient.Do made.