Shuffle knative.dev/pkg/tracing code a bit (#581)
This moves around the code within the tracing package in preparation for adding support for other exporter options.

As the fields changed are all private, this should be a no-op for consumers of this package.

The goal of this change is to make it so that no files mention zipkin other than zipkin.go (or tests, which I left alone).