Add a Mako stub sidecar
The stub sidecar replaces the Mako sidecar for contributors that can't
access the Mako service. It emits recorded data as CSV to stdout.

The first line is a comment containing a JSON dump of BenchmarkInfo.
The last line is a comment containing the column names.

Each line in between is a CSV row. The first two columns are always
inputValue and errorMessage, then one column per metric value.

Run Aggregate data points have an empty inputValue.

Example:

    # {"benchmarkKey":"5903682180743168","tags":["runOwner=grantrodgers","project-id=grantrodgers-crd","commit=5102ef7","kubernetes=v1_13_6-gke_13","go1_12_5"]}
    1566927060345.571777,,0.039124
    1566927060346.181152,,0.046175
    1566926807393.443115,,,16.525982
    ,,,,0.000000
    ,,,,,7.000000
    1566926807388.952393,Timeout waiting for delivery,,,,
    # inputValue,errorMessage,pl,dl,pe,de

The proto files are copied from google/mako because they're in an
"internal" path and those can't be imported.