Fix spoof to not always append port
In #630 we changed transport spoofer behavior to always
append a port to our resolved host. A valid use case is to
supply a full l4 dest as the spoof targed (myhost:port). In
this case we now append an extra port resulting in an invalid
dest. Lets keep our original behavior where assume our spoofed
endpoint is a l4 dest, not a hostname.