Add glog/klog->Zap redirect designs to repo
It is not possible to test updating the Gopkg.toml to override usage
of glog or klog to these libraries until they are already in the repo.

logging.go has been updated to supply the value logging.Verbosity, but
it's not intended to be the final way it is setup.