Make sure we actually check the last value.
Currently the function aptly named CheckLastValueData does not only check that, but
also checks that it's the _single_ entry, which is not necessarily true,
especially if run as part of a bigger test suite.
So change it to actually verify the last row and remove the check for the singular entry.
Given that all the tests pass, this probably was not intented anyway.