Add BucketSet type which implements consistent bucketing (#1477)
* Add BucketSet type which implements consistent bucketing

BucketSet permits mapping keys to buckets in a consistent manner, permitting us to use this type
in the HA applications. BucketSet implements the Bucket interface so it can be used in reconciler/leaderelection
pieces and has an additional method which returns the owner for the key. This is necessary in autoscaler
where requests need to be forwarded to an appropriate autoscaler.

* git mod stuff

* fixes

* nit

* two types for two thigns

* more tests