Allow passing custom cmp.Options to TableTest. (#1552)
We're playing around with writing TableTests for the Operator (which uses manifestival) and would
like to see if we can get a good comparator for the automatically added last-applied-configuration annotation.
If that ends up being general purpose we can upstream it here, but in general it's nice to be allowed to do
some application specific filtering of the desired state changes.