Permit table test deletions to arrive in arbitrary order.
If you haver ever written a table test, perhaps you noticed that when several objects
are deleted you need to list them is a particular order. And if you're deleting things in parallel
it's just impossible.
But what we care is that N objects are deleted -- so verify that.
The key is type-name-ns, which should work most of the time.

I've cleaned the tests in Eventing to work with this change
and there's an open PR in serving that would make the tests pass
with this change there too

Fixes #411