Teach the spoofing client to support NodePort.
This is loosely based on similar logic that I added to kingress conformance testing here:
https://github.com/knative/networking/blob/611f989aa593e513688d0f0047701e8352571ac4/test/conformance/ingress/util.go#L904-L912

The basic idea is that in our custom dialer we look at the requested port and translate that to the service's NodePort when present.

This is in support of running Serving's e2e testing on top of KinD environments without service `type: LoadBalancer`.