Delay start of informers in sharedmain (#1767) (#1838)
The EnableInjectionOrDie starts informers immedidately in a go
routine, which changes the order of informer start and controller
setup. As a result setting resync period to informer handler will
be ignored as the informer is already started. Though this will
likely affect more components, changing the EnableInjectionOrDie
function will break the API, so this is a temporary solution to
fix all the controllers depending on sharedmain.

Co-authored-by: Zhongduo Lin (Jimmy) <zhongduo@google.com>