Add a test package for making assertions about dependencies.
See the example test for illustrative examples of usage.  It is notable that the gofuzz test failed until I rebased on Ville's recent change.

The goal here is ultimately to assert that our key libraries don't take on heavy dependencies to combat the binary-size bloat that currently afflicts the queue-proxy.