Abstract how our webhook implementations target things.
This change introduces a `targeter.Interface` abstraction that we've been
using downstream to abstract the core logic of webhooks from the details
of how delivery to them is handled.

The most immediate benefit of this refactoring is that each of our webhooks
no longer needs direct knowledge of `caBundle` and the secret logic, which
is consolidated in the provided implementations.  So for examples, folks
could now provide an implementation of `targeter.Interface` that receives
its `caBundle` information in a wholly different way (e.g. mounted from the
filesystem).

This encapsulation of how the `WebhookClientConfig` is produced also enables
folks to produce these configurations in more imaginative ways. For example,
I have included an implementation that allows folks to host a webhook with a
`BasePath()`, produce configs for "names" under this `BasePath()` and
then extract the "name" being handled when receiving a webhook request.

This encapsulation also lets folks produce `WebhookClientConfig` blocks that
don't even need `caBundle` and leverage `URL` blocks, if they are hosting
things with public-CA TLS-terminated endpoints (e.g. as a `ksvc`).