Multi-controller leadership electors
This is an approach to constructing and running one leadership elector
that promotes and demotes more than one controller/reconciler in tandem
based on a single lease. Such capability is a must when it is desirable
for several related reconcilers to run from the same process.

Relates to https://github.com/knative/serving/issues/13447